[YS-301] refactor: 사용자 정보 헤더 컴포넌트 분리
