+++ src/app/user/profile/components/ParticipantUserInfo/ParticipantUserInfo.css.ts
@@ -11,40 +11,6 @@ export const joinLayout = style({
   flexGrow: 1,
 });
 
-export const headerContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '0.8rem',
-});
-
-export const titleContainer = style({
-  display: 'flex',
-  alignItems: 'center',
-  gap: '0.8rem',
-});
-
-export const badge = style({
-  ...fonts.label.large.SB14,
-  borderRadius: '1.2rem',
-  padding: '0.4rem 0.8rem',
-  backgroundColor: colors.primaryTinted,
-  color: colors.primaryMint,
-});
-
-export const title = style({
-  ...fonts.title.medium.SB20,
-  color: colors.text06,
-});
-
-export const infoContainer = style({
-  ...fonts.body.normal.R16,
-  color: colors.text05,
-  display: 'flex',
-  alignItems: 'center',
-  gap: '1.2rem',
-  padding: '0 0.8rem',
-});
-
 export const updateInfoFormContainer = style({
   display: 'flex',
   flexDirection: 'column',
@@ -61,32 +27,6 @@ export const updateInfoForm = style({
   backgroundColor: colors.field03,
   borderRadius: '1.2rem',
   padding: '2.4rem 4rem',
-});
-
-export const verticalLine = style({
-  position: 'relative',
-  height: '100%',
-  selectors: {
-    '&::after': {
-      content: '""',
-      width: '0.1rem',
-      height: '1.6rem',
-      position: 'absolute',
-      top: '50%',
-      right: '0',
-      transform: 'translateY(-50%)',
-      backgroundColor: colors.line03,
-    },
-    '&:last-child::after': {
-      display: 'none',
-    },
-  },
-});
-
-export const emailWrapper = style({
-  display: 'flex',
-  alignItems: 'center',
-  gap: '0.8rem',
 });
 
 export const updateButton = style({
