+++ src/app/user/profile/page.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import ParticipantUserInfo from './components/ParticipantUserInfo/ParticipantUserInfo';
+import UserInfoHeader from './components/UserInfoHeader/UserInfoHeader';
 import { joinLayout } from './ProfilePage.css';
 
 import useUserInfo from '@/app/home/hooks/useUserInfo';
@@ -14,6 +15,7 @@ const ProfilePage = () => {
   if (isParticipantInfo(userInfo)) {
     return (
       <section className={joinLayout}>
+        <UserInfoHeader userInfo={userInfo} />
         <ParticipantUserInfo userInfo={userInfo} />
       </section>
     );
